requirements.txt
fastapi
langchain
langchain-google-genai
langchain_core
uvicorn[standard]
python-dotenv
langchain_community
langgraph
langchain-openai
openai